pyproject.toml
[project]
name = "qrcode-generator"
version = "0.1.0"
description = "Cross-platform QR code generator with CLI and GUI"
authors = [
    { name="Pierre Lacerte" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pillow==11.2.1",
    "qrcode==8.2",
]

[project.optional-dependencies]
dev = [
    "pyinstaller==6.6.0",
]

[project.scripts]
qrcode = "qrcode_generator.qr:main"

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
